List with all products rendered in CategoriesAndCatalog component. Catalog component deleted
